WIP integration test fixes

+++ py/desispec/test/integration_test.py
@@ -101,6 +101,26 @@ def sim(night, nspec=5, clobber=False):
 
     return
 
+def run_pipeline_step(tasktype):
+    """Convenience wrapper to run a pipeline step"""
+    #- First count the number of tasks that are ready
+    log = logging.get_logger()
+
+    dbpath = io.get_pipe_database()
+    db = pipe.load_db(dbpath, mode="r")
+    task_count = db.count_task_states(tasktype)
+    count_string = ', '.join(['{:2d} {}'.format(x[1], x[0]) for x in task_count.items()])
+
+    nready = task_count['ready']
+    if nready > 0:
+        log.info('{:16s}: {}'.format(tasktype, count_string))
+        com = "desi_pipe tasks --tasktype {tasktype} | grep -v DEBUG | desi_pipe script --tasktype {tasktype}".format(tasktype=tasktype)
+        log.info('Running {}'.format(com))
+        script = sp.check_output(com, shell=True)
+        log.info('Running {}'.format(script))
+        sp.check_call(script, shell=True)
+    else:
+        log.warning('{:16s}: {} -- SKIPPING'.format(tasktype, count_string))
 
 def integration_test(night=None, nspec=5, clobber=False):
     """Run an integration test from raw data simulations through redshifts
@@ -115,12 +135,9 @@ def integration_test(night=None, nspec=5, clobber=False):
 
     """
     log = logging.get_logger()
-    log.setLevel(logging.DEBUG)
 
     # YEARMMDD string, rolls over at noon not midnight
-    # TODO: fix usage of night to be something other than today
     if night is None:
-        #night = time.strftime('%Y%m%d', time.localtime(time.time()-12*3600))
         night = "20160726"
 
     # check for required environment variables
@@ -136,11 +153,16 @@ def integration_test(night=None, nspec=5, clobber=False):
 
     # create production
 
-    if os.path.isdir(proddir):
+    if clobber and os.path.isdir(proddir):
         shutil.rmtree(proddir)
 
-    com = "desi_pipe create --db-sqlite"
-    sp.check_call(com, shell=True)
+    dbfile = io.get_pipe_database()
+    if not os.path.exists(dbfile):
+        com = "desi_pipe create --db-sqlite"
+        log.info('Running {}'.format(com))
+        sp.check_call(com, shell=True)
+    else:
+        log.info("Using pre-existing production database {}".format(dbfile))
 
     # Modify options file to restrict the spectral range
 
@@ -152,89 +174,10 @@ def integration_test(night=None, nspec=5, clobber=False):
     opts['psf']['nspec'] = nspec
     pipe.prod.yaml_write(optpath, opts)
 
-    # Run preprocessing
-
-    com = "desi_pipe tasks --tasktype pix | grep -v DEBUG | desi_pipe script --tasktype pix"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run psf estimation
-
-    com = "desi_pipe tasks --tasktype psf | grep -v DEBUG | desi_pipe script --tasktype psf"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run psfnight combine
-
-    com = "desi_pipe tasks --tasktype psfnight | grep -v DEBUG | desi_pipe script --tasktype psfnight"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run extraction
-
-    com = "desi_pipe tasks --tasktype extract | grep -v DEBUG | desi_pipe script --tasktype extract"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run fiberflat
-
-    com = "desi_pipe tasks --tasktype fiberflat | grep -v DEBUG | desi_pipe script --tasktype fiberflat"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run fiberflat night
-
-    com = "desi_pipe tasks --tasktype fiberflatnight | grep -v DEBUG | desi_pipe script --tasktype fiberflatnight"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run sky
-
-    com = "desi_pipe tasks --tasktype sky | grep -v DEBUG | desi_pipe script --tasktype sky"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run starfit
-
-    com = "desi_pipe tasks --tasktype starfit | grep -v DEBUG | desi_pipe script --tasktype starfit"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run fluxcalib
-
-    com = "desi_pipe tasks --tasktype fluxcalib | grep -v DEBUG | desi_pipe script --tasktype fluxcalib"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run cframe
-
-    com = "desi_pipe tasks --tasktype cframe | grep -v DEBUG | desi_pipe script --tasktype cframe"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run spectra
-
-    com = "desi_pipe tasks --tasktype spectra | grep -v DEBUG | desi_pipe script --tasktype spectra"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
-
-    # Run redshifts
-
-    com = "desi_pipe tasks --tasktype redshift | grep -v DEBUG | desi_pipe script --tasktype redshift"
-    script = sp.check_output(com, shell=True)
-    print("running {}".format(script))
-    sp.check_call(script, shell=True)
+    # Run the pipeline tasks in order
+    from desispec.pipeline.tasks.base import default_task_chain
+    for tasktype in default_task_chain:
+        run_pipeline_step(tasktype)
 
     # #-----
     # #- Did it work?
@@ -250,12 +193,27 @@ def integration_test(night=None, nspec=5, clobber=False):
     nside=64
     pixels = np.unique(radec2pix(nside, fibermap['RA_TARGET'], fibermap['DEC_TARGET']))
 
+    num_missing = 0
+    for pix in pixels:
+        zfile = io.findfile('zbest', groupname=pix)
+        if not os.path.exists(zfile):
+            log.error('Missing {}'.format(zfile))
+            num_missing += 1
+
+    if num_missing > 0:
+        log.critical('{} zbest files missing'.format(num_missing))
+        sys.exit(1)
+
     print()
     print("--------------------------------------------------")
     print("Pixel     True  z        ->  Class  z        zwarn")
     # print("3338p190  SKY   0.00000  ->  QSO    1.60853   12   - ok")
     for pix in pixels:
         zfile = io.findfile('zbest', groupname=pix)
+        if not os.path.exists(zfile):
+            log.error('Missing {}'.format(zfile))
+            continue
+
         zfx = fits.open(zfile, memmap=False)
         zbest = zfx['ZBEST'].data
         for i in range(len(zbest['Z'])):
